fixed ucs module (Bug #14882)
git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.0/ucs-school@30322 1b283449-9f2f-0410-b571-8eb228e1a901

+++ ucs-school-netlogon/conffiles/ucs-school-netlogon.py
@@ -85,7 +85,7 @@ def handler(configRegistry, changes):
 
 	# delete old ucsschool/import/set/netlogon/script/path
 	old = changes.get("ucsschool/import/set/netlogon/script/path", "")
-	if len(old) > 0 and old[0]:
+	if old and len(old) > 0 and old[0]:
 		old = os.path.join(netlogonDir, old[0])
 		if os.path.isfile(old):
 			os.remove(old)
